pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lakecube"
version = "0.0.1"
description = "A thin cube contract over Databricks — modern Essbase, re-engineered on the Lakehouse."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "Lakecube contributors" },
]
keywords = ["olap", "essbase", "databricks", "semantic-layer", "cube", "epm", "finance"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]
databricks = [
    "databricks-sdk>=0.20",
]

[project.scripts]
lakecube = "lakecube.cli.main:cli"

[project.urls]
Homepage = "https://github.com/Tushar1344/lakecube"
Issues = "https://github.com/Tushar1344/lakecube/issues"

[tool.hatch.build.targets.wheel]
packages = ["lakecube"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
